+++ tests/Helpers/TestCase.php
@@ -5,10 +5,42 @@ namespace Popstas\Transmission\Console\Tests\Helpers;
 abstract class TestCase extends \PHPUnit_Framework_TestCase
 {
     public $expectedTorrentList = [
-        ['downloadDir' => '/', 'id' => 1, 'name' => 'name.ext', 'totalSize' => 1, 'uploadedEver' => 3],
-        ['downloadDir' => '/', 'id' => 2, 'name' => 'name.ext', 'totalSize' => 2, 'uploadedEver' => 4],
-        ['downloadDir' => '/', 'id' => 3, 'name' => 'name2.ext', 'totalSize' => 3, 'uploadedEver' => 5],
-        ['downloadDir' => '/dir2', 'id' => 4, 'name' => 'name.ext', 'totalSize' => 1, 'uploadedEver' => 5]
+        [
+            'downloadDir'  => '/',
+            'id'           => 1,
+            'name'         => 'name.ext',
+            'totalSize'    => 1,
+            'uploadedEver' => 3,
+            'downloadedEver' => 1,
+            'doneDate' => 1456784850,
+        ],
+        [
+            'downloadDir'  => '/',
+            'id'           => 2,
+            'name'         => 'name.ext',
+            'totalSize'    => 2,
+            'uploadedEver' => 4,
+            'downloadedEver' => 1,
+            'doneDate' => 1456784850,
+        ],
+        [
+            'downloadDir'  => '/',
+            'id'           => 3,
+            'name'         => 'name2.ext',
+            'totalSize'    => 3,
+            'uploadedEver' => 5,
+            'downloadedEver' => 1,
+            'doneDate' => 1456784850,
+        ],
+        [
+            'downloadDir'  => '/dir2',
+            'id'           => 4,
+            'name'         => 'name.ext',
+            'totalSize'    => 1,
+            'uploadedEver' => 5,
+            'downloadedEver' => 1,
+            'doneDate' => 0,
+        ],
     ];
 
     /**